+++ jslib/mappanel/lib/mappanel.js
@@ -11,7 +11,8 @@ define('mappanel', function (require, exports, module) {
     , SPLITTER = /[\r\n]+/g
     , CR = '\r'
     , geolocation = window.navigator.geolocation
-    , $win = $(window);
+    , $win = $(window)
+    , isIE = $.browser.msie;
 
   var Panel = require('panel');
 
@@ -28,6 +29,7 @@ define('mappanel', function (require, exports, module) {
                 + '<div class="map-place">'
                   + '<div class="place-editor">'
                     + '<i class="pointer icon24-enter place-submit"></i>'
+                    + '<div class="place-filter"></div>'
                     + '<textarea class="normal" name="place-text" id="place-text" placeholder="Enter place here."></textarea>'
                   + '</div>'
                   + '<div class="map-places hide">'
@@ -197,7 +199,8 @@ define('mappanel', function (require, exports, module) {
 
     , change: function (place, searchable) {
         var placeData = this.place
-          , oldTitle = placeData.title;
+          , oldTitle = placeData.title
+          , oldDesc = placeData.description;
         placeData.title = place.title;
         placeData.description = place.description;
         placeData.lat = place.lat || '';
@@ -210,7 +213,9 @@ define('mappanel', function (require, exports, module) {
         } else {
           this.placeInput.change(printPlace(place.title, place.description));
         }
-        this.emit('update-place', placeData);
+        if (oldTitle !== place.title || oldDesc !== place.description) {
+          this.emit('update-place', placeData);
+        }
       }
 
     , revert: function (e) {
@@ -334,8 +339,8 @@ define('mappanel', function (require, exports, module) {
 
     , keyup: function (e) {
         switch (e.keyCode) {
-          case 40: // down arrow
-          case 38: // up arrow
+          case 40: // down
+          case 38: // up
           case 16: // shift
           case 17: // ctrl
           case 18: // alt
@@ -360,6 +365,17 @@ define('mappanel', function (require, exports, module) {
             e.preventDefault();
             component.emit('placeinput-tab');
             break;
+          case 40: // down
+            // if the cursor in the last, tab to PlacesList
+            var v = this.$element.val()
+              , l = v.length
+              , ele = this.$element[0]
+              , end = selectionEnd(ele);
+            if (l === end) {
+              e.preventDefault();
+              component.emit('placeinput-tab');
+            }
+            break;
         }
       }
 
@@ -371,7 +387,7 @@ define('mappanel', function (require, exports, module) {
       }
 
     , keydown: function (e) {
-        this.suppressKeyPressRepeat = !!~R.indexOf([9], e.keyCode);
+        this.suppressKeyPressRepeat = !!~R.indexOf([9, 40], e.keyCode);
         this.keyHandler(e);
       }
 
@@ -670,14 +686,6 @@ define('mappanel', function (require, exports, module) {
 
           this.createIcons();
 
-          this.disableOptions = {
-              panControl: false
-            , zoomControl: false
-            , scaleControl: false
-            , mapTypeControl: false
-            , overviewMapControl: false
-          };
-
           this.enableOptions = {
               //zoom: this.zoomNum
             //, panControl: true
@@ -699,11 +707,10 @@ define('mappanel', function (require, exports, module) {
             , {
                 zoom: this.zoomNum
               , center: this._center
+              , disableDefaultUI: true
               , MapTypeId: GMaps.MapTypeId.ROADMAP
             }
           );
-
-          this._map.setOptions(this.disableOptions);
 
           this._overlay = new GMaps.OverlayView();
           this._overlay.draw = function () {};
@@ -938,7 +945,7 @@ define('mappanel', function (require, exports, module) {
     placeString || (placeString = '');
     var ps = placeString.split(SPLITTER)
       , title = ps.length ? $.trim(ps.shift()) : ''
-      , description = $.trim(ps.join(CR));
+      , description = $.trim(ps.join(CR)).replace(SPLITTER, '');
 
     return {
         title: title
@@ -949,6 +956,20 @@ define('mappanel', function (require, exports, module) {
   var printPlace = function (title, description) {
     return title + (description ? CR + description.replace(SPLITTER, CR) : '');
   };
+
+  // get Textarea selectionEnd
+  var selectionEnd = function (inputor) {
+    return isIE ? getIESelectionEnd(inputor) : inputor.selectionEnd;
+  }
+
+  var getIESelectionEnd = function (inputor) {
+    var r = document.selection.createRange()
+      , re = inputor.createTextRange()
+      , rc = re.duplicate();
+    re.moveToBookmark(r.getBookmark());
+    rc.setEndPoint('EndToStart', re);
+    return rc.text.length + r.text.length;
+  }
 
   return MapPanel;
 });